pyproject.toml
[project]
name = "enterprise-rag"
version = "1.0.0"
description = "Enterprise Multimodal Hybrid RAG Platform"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # --- FastAPI & Server ---
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
    # --- Pydantic ---
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    # --- Vector DB ---
    "qdrant-client>=1.10.0",
    # --- GCP Services ---
    "google-cloud-firestore>=2.16.0",
    "google-cloud-storage>=2.17.0",
    "google-auth>=2.30.0",
    "google-auth-oauthlib>=1.2.0",
    # --- Auth ---
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0",
    # --- LangChain / LangGraph ---
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.2.0",
    # --- PDF Processing (Docling) ---
    "docling>=2.0.0",
    "docling-core>=2.0.0",
    # --- Monitoring ---
    "langfuse>=2.40.0",
    "prometheus-client>=0.21.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    # --- Evaluation ---
    "ragas>=0.2.0",
    # --- Logging ---
    "structlog>=24.2.0",
    # --- Utilities ---
    "tenacity>=8.3.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "black>=24.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

# Legacy PDF parsers (optional — replaced by docling)
legacy-pdf = ["PyMuPDF>=1.24.0", "pdfplumber>=0.11.0"]

# In-process ML models (optional — replaced by Ollama)
local-models = [
    "sentence-transformers>=3.0.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "pillow>=10.3.0",
    "einops>=0.8.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.black]
target-version = ["py311"]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests (no external services)",
    "integration: Integration tests (require services)",
]
